Extract common test code between ApiGateway and HttpIntegration into base class
